requirements.txt
fastapi
uvicorn
chromadb
openai
python-dotenv
pydantic
pytest
httpx